feat: clean up run function
