+++ src/experiment.ts
@@ -7,7 +7,6 @@
  */
 // You can import stylesheets (.scss or .css).
 // Import required plugins and modules from jsPsych
-import FullscreenPlugin from '@jspsych/plugin-fullscreen';
 import jsPsychHtmlKeyboardResponse from '@jspsych/plugin-html-keyboard-response';
 import PreloadPlugin from '@jspsych/plugin-preload';
 import jsPsychSurveyHtmlForm from '@jspsych/plugin-survey-html-form';
@@ -18,8 +17,12 @@ import { JsPsych, initJsPsych } from 'jspsych';
 import '../styles/main.scss';
 import { groupInstructions, tipScreen } from './instructions';
 import { showEndScreen } from './quit';
-import { quitBtnAction } from './quit';
-import { DeviceConnectPages, generatePreloadStrings, resize } from './setup';
+import {
+  DeviceConnectPages,
+  fullScreenPlugin,
+  generatePreloadStrings,
+  resize,
+} from './setup';
 import {
   connectToSerial,
   connectToUSB,
@@ -262,6 +265,7 @@ export async function run({
   const blocks_per_half: number = 5;
   const device_name: string = 'Arduino Micro';
 
+  //Pseudo state variable
   let device_info: {
     device: (SerialPort & USBDevice) | null;
     send_trigger_func: (
@@ -276,6 +280,7 @@ export async function run({
     ) => {},
   };
 
+  //initialize jspsych
   const jsPsych: JsPsych = initJsPsych({
     show_progress_bar: true,
     auto_update_progress_bar: false,
@@ -285,8 +290,14 @@ export async function run({
     },
   });
 
+  // Randomize order of countables
+  let exp_parts_cntables: ('people' | 'objects')[] = ['people', 'objects'];
+  exp_parts_cntables = jsPsych.randomization.shuffle(exp_parts_cntables);
+
+  //initiate timeline
   const timeline: timeline = [];
 
+  //push trials to timeline
   jsPsych;
   // Preload assets
   timeline.push({
@@ -299,38 +310,10 @@ export async function run({
     DeviceConnectPages(jsPsych, device_info, connectToUSB, device_name),
   );
 
-  // Switch to fullscreen
-  timeline.push({
-    type: FullscreenPlugin,
-    fullscreen_mode: true,
-    message: '',
-    button_label: i18next.t('fullscreen'),
-    on_load: function (): void {
-      const quit_btn: HTMLButtonElement = document.createElement('button');
-      quit_btn.type = 'button';
-      quit_btn.setAttribute(
-        'style',
-        'color: #fff; border-radius: 4px; background-color: #1d2124; border-color: #171a1d; position: absolute; right: 1%; top: 50%; transform: translateY(-50%)',
-      );
-
-      quit_btn.addEventListener('click', () => quitBtnAction(jsPsych));
-
-      quit_btn.appendChild(document.createTextNode(i18next.t('quitBtn')));
-
-      document
-        .getElementById('jspsych-progressbar-container')!
-        .appendChild(quit_btn);
-    },
-  });
-
-  timeline.push(resize(jsPsych));
-
-  // Randomize order of countables
-  let exp_parts_cntables: ('people' | 'objects')[] = ['people', 'objects'];
-  exp_parts_cntables = jsPsych.randomization.shuffle(exp_parts_cntables);
-
   // Run numerosity task
   timeline.push(
+    fullScreenPlugin(jsPsych),
+    resize(jsPsych),
     groupInstructions(jsPsych, exp_parts_cntables[0]),
     tipScreen(),
     createButtonPage(
