+++ src/setup.ts
@@ -1,8 +1,10 @@
+import FullscreenPlugin from '@jspsych/plugin-fullscreen';
 import HtmlButtonResponsePlugin from '@jspsych/plugin-html-button-response';
 import JsResize from '@jspsych/plugin-resize';
 import i18next from 'i18next';
 import { DataCollection, JsPsych } from 'jspsych';
 
+import { quitBtnAction } from './quit';
 import {
   connectToSerial,
   connectToUSB,
@@ -224,3 +226,28 @@ export function DeviceConnectPages(
     },
   };
 }
+
+export const fullScreenPlugin: (jsPsych: JsPsych) => timeline = (
+  jsPsych: JsPsych,
+) => ({
+  type: FullscreenPlugin,
+  fullscreen_mode: true,
+  message: '',
+  button_label: i18next.t('fullscreen'),
+  on_load: function (): void {
+    const quit_btn: HTMLButtonElement = document.createElement('button');
+    quit_btn.type = 'button';
+    quit_btn.setAttribute(
+      'style',
+      'color: #fff; border-radius: 4px; background-color: #1d2124; border-color: #171a1d; position: absolute; right: 1%; top: 50%; transform: translateY(-50%)',
+    );
+
+    quit_btn.addEventListener('click', () => quitBtnAction(jsPsych));
+
+    quit_btn.appendChild(document.createTextNode(i18next.t('quitBtn')));
+
+    document
+      .getElementById('jspsych-progressbar-container')!
+      .appendChild(quit_btn);
+  },
+});
